requirements.txt
pyjwt~=2.3.0
Flask~=2.0.2
sqlalchemy~=1.4.31
flask-sqlalchemy~=2.5.1
psycopg2-binary~=2.9.3
